Added Login File Timestamp
Login now creates or appends a file name LoginFile.txt when the user successfully logs into the application. This file will containt the users name and the local timestamp of when the login occurred.

Also added a tab control to the main screens form to switch between appointment views and customer views.